Moved ViewModels from DAL to DTOs and renamed them
