+++ TransIT.BLL/DTOs/ChangePasswordDTO.cs
@@ -1,0 +1,7 @@
+﻿namespace TransIT.BLL.DTOs
+{
+    public class ChangePasswordDTO
+    {
+        public string Password { get; set; }
+    }
+}
