+++ TransIT.BLL/DTOs/LoginDTO.cs
@@ -1,6 +1,6 @@
-namespace TransIT.DAL.Models.ViewModels
+namespace TransIT.BLL.DTOs
 {
-    public class LoginViewModel
+    public class LoginDTO
     {
         public string Login { get; set; }
         public string Password { get; set; }
