--- TransIT.DAL/Models/ViewModels/ChangePasswordViewModel.cs
@@ -1,9 +1,0 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace TransIT.DAL.Models.ViewModels
-{
-    public class ChangePasswordViewModel
-    {
-        public string Password { get; set; }
-    }
-}
